pyproject.toml
[project]
name = "practice-fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.123.0",
    "imagekitio>=4.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
